LayoutTests/platform/ios/imported/w3c/web-platform-tests/css/css-flexbox/flex-item-compressible-001-expected.txt

Test1: "width: 100%"






Test2: "width: calc(100%)"






Test3: "width: calc(140px + 100%)"







PASS .flexbox 1
PASS .flexbox 2
FAIL .flexbox 3 assert_equals:
<div class="flexbox">
      <div class="spacer"></div>
      <input type="button" value="XXXXXXX" class="test1" data-expected-width="140">
    </div>
width expected 140 but got 180
FAIL .flexbox 4 assert_equals:
<div class="flexbox">
      <div class="spacer"></div>
      <input type="submit" value="XXXXXXX" class="test1" data-expected-width="140">
    </div>
width expected 140 but got 180
FAIL .flexbox 5 assert_equals:
<div class="flexbox">
      <div class="spacer"></div>
      <input type="reset" value="XXXXXXX" class="test1" data-expected-width="140">
    </div>
width expected 140 but got 180
PASS .flexbox 6
PASS .flexbox 7
FAIL .flexbox 8 assert_equals:
<div class="flexbox">
      <div class="spacer"></div>
      <input type="button" value="XXXXXXX" class="test2" data-expected-width="140">
    </div>
width expected 140 but got 180
FAIL .flexbox 9 assert_equals:
<div class="flexbox">
      <div class="spacer"></div>
      <input type="submit" value="XXXXXXX" class="test2" data-expected-width="140">
    </div>
width expected 140 but got 180
FAIL .flexbox 10 assert_equals:
<div class="flexbox">
      <div class="spacer"></div>
      <input type="reset" value="XXXXXXX" class="test2" data-expected-width="140">
    </div>
width expected 140 but got 180
FAIL .flexbox 11 assert_equals:
<div class="flexbox">
      <div class="spacer"></div>
      <input type="text" class="test3" data-expected-width="140">
    </div>
width expected 140 but got 100
FAIL .flexbox 12 assert_equals:
<div class="flexbox">
      <div class="spacer"></div>
      <input type="range" class="test3" data-expected-width="140">
    </div>
width expected 140 but got 100
FAIL .flexbox 13 assert_equals:
<div class="flexbox">
      <div class="spacer"></div>
      <input type="button" value="XXXXXXX" class="test3" data-expected-width="140">
    </div>
width expected 140 but got 180
FAIL .flexbox 14 assert_equals:
<div class="flexbox">
      <div class="spacer"></div>
      <input type="submit" value="XXXXXXX" class="test3" data-expected-width="140">
    </div>
width expected 140 but got 180
FAIL .flexbox 15 assert_equals:
<div class="flexbox">
      <div class="spacer"></div>
      <input type="reset" value="XXXXXXX" class="test3" data-expected-width="140">
    </div>
width expected 140 but got 180
